Adjusted Draw Weapon and Light/Heavy Attack Inputs
Changed attack functions in Base so that pressing any of the attack weapons only draw the weapon instead of drawing and then attacking, because the latter felt awkward.

Added MovementOn variable to Base class and Movement setter function to CharacterActions class.
Used notifiers in Light/Heavy Attack 1 montages and Animation Blueprint to set it so that the player doesn't attack and move at the same time.